#4035 Added support for warm start of Ipopt. A boolean "self.warm_start" inside the collocation algorithm keeps track of wheter warm start are to be used or not. The warm start options are set from outside the collocation class, in this case from the MPC-class.
#4036
Substituted the "_eliminate_fixed_parameters" in the collocation algorithm 
where the parameter symbols were substituted for their value for a new method
"_create_nlp_parameters" where the parameter symbols are substituted for new
parameter symbols "self.pp = MX(par)". The parameter values are stored in a 
vector "self._par_vals".  

#4047
Created a class which handles all of the MPC stuff. When initiated, an instance
of a LocalDAECollocator-object is created. The MPC-class keeps this object
and is able to update it with measurements of the states (or with simulated 
noise) before each optimization. 

The MPC-class also:
Redefines the initial trajectories before each optimization (based on the 
results from the previous optimization).
Defines the warm start options for the Ipopt, aswell as blocking factors 
for the input.
Measures the time for the different tasks: initializing, updating, solving and 
post-processing. The maximum total time for one sample is also saved.

A new result class "MPCResult" has also been added to jmi_algorithm_drivers. 
The result class prints the times of the MPC-optimization and returns a 
resultobject to the user.
         


git-svn-id: https://svn.jmodelica.org/trunk@6911 7ae51c65-13d5-4921-9d80-6511ddd1419c