Account for more of the time used for optimization again after the restructuring in r7240, for #4168.
Note: Before, time spent in `solve_nlp` and not accounted for in `sol_time` would be accounted for as `init` time.
Now, it will be accounted for as `preprocessing_time` instead, since `init` time should not change when reoptimizing.
Perhaps we should introduce another timing category for initialization that occurs per optimization.

git-svn-id: https://svn.jmodelica.org/trunk@7244 7ae51c65-13d5-4921-9d80-6511ddd1419c