#5126
A workaround was implemented as delegators are created as static fields and thus the current OptionRegistry could be invalid upon C-code compilation. Instead it is up to the respective delegators to utilize the OptionRegistry upon compilation. As such the way in which C-compiler path and version numbers are calculated is dependent on the current OptionRegistry and the implementations might differ; compilerVersionNumber is not required for each delegator as the method signature could vary.

git-svn-id: https://svn.jmodelica.org/trunk@9243 7ae51c65-13d5-4921-9d80-6511ddd1419c